requirements.txt
flask
dnspython
